requirements.txt
pyshark>=0.6.0
pyyaml>=6.0.1
orjson>=3.9.15
rich>=13.7.1